+++ app/process-1/page.tsx
@@ -1,10 +1,5 @@
-import { DashboardShell } from "@/components/dashboard/DashboardShell";
-import { Process1Workbench } from "@/components/dashboard/process1/Process1Workbench";
+import { redirect } from "next/navigation";
 
 export default function Process1Page() {
-  return (
-    <DashboardShell>
-      <Process1Workbench />
-    </DashboardShell>
-  );
+  redirect("/");
 }